[config] a little cleanup
This is just reordering a few fields and merging secure storage helpers

This was enabled by moving toward yaml

Closes: #4642
Approved by: mimoo
