+++ config/src/config/debug_interface_config.rs
@@ -9,20 +9,19 @@ use std::collections::HashMap;
 #[serde(default, deny_unknown_fields)]
 pub struct DebugInterfaceConfig {
     pub admission_control_node_debug_port: u16,
-    // This has similar use to the core-node-debug-server itself
-    pub metrics_server_port: u16,
-    pub public_metrics_server_port: u16,
     pub address: String,
     pub libra_trace: LibraTraceConfig,
+    pub metrics_server_port: u16,
+    pub public_metrics_server_port: u16,
 }
 
 impl Default for DebugInterfaceConfig {
     fn default() -> DebugInterfaceConfig {
         DebugInterfaceConfig {
             admission_control_node_debug_port: 6191,
+            address: "0.0.0.0".to_string(),
             metrics_server_port: 9101,
             public_metrics_server_port: 9102,
-            address: "0.0.0.0".to_string(),
             libra_trace: LibraTraceConfig::default(),
         }
     }
