+++ config/src/config/logger_config.rs
@@ -8,10 +8,10 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 #[serde(default, deny_unknown_fields)]
 pub struct LoggerConfig {
-    // Use async logging
-    pub is_async: bool,
     // channel size for the asychronous channel for node logging.
     pub chan_size: usize,
+    // Use async logging
+    pub is_async: bool,
     // The default logging level for slog.
     pub level: Level,
 }
@@ -19,8 +19,8 @@ pub struct LoggerConfig {
 impl Default for LoggerConfig {
     fn default() -> LoggerConfig {
         LoggerConfig {
-            is_async: true,
             chan_size: CHANNEL_SIZE,
+            is_async: true,
             level: Level::Info,
         }
     }
