+++ config/src/config/mempool_config.rs
@@ -7,15 +7,14 @@ use serde::{Deserialize, Serialize};
 #[serde(default, deny_unknown_fields)]
 pub struct MempoolConfig {
     pub broadcast_transactions: bool,
-    pub shared_mempool_tick_interval_ms: u64,
+    pub capacity: usize,
+    pub capacity_per_user: usize,
+    pub max_broadcasts_per_peer: usize,
     pub shared_mempool_backoff_interval_ms: u64,
     pub shared_mempool_batch_size: usize,
     pub shared_mempool_max_concurrent_inbound_syncs: usize,
     pub shared_mempool_min_broadcast_recipient_count: Option<usize>,
-    pub max_broadcasts_per_peer: usize,
-    pub capacity: usize,
-    // max number of transactions per user in Mempool
-    pub capacity_per_user: usize,
+    pub shared_mempool_tick_interval_ms: u64,
     pub system_transaction_timeout_secs: u64,
     pub system_transaction_gc_interval_ms: u64,
 }
